#ELLIE-102 : 로드 매니저 이름 수정
1. SceneCenter -> SceneLoadManager
