+++ Assets/Resources/Sounds/ellie_sound1.wav.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ea2afe5f9f47af649a3db786d40fa1b1
+guid: 3258489f1f6a9a7478e534f8b377ec83
 AudioImporter:
   externalObjects: {}
   serializedVersion: 6
