+++ Assets/Resources/Sounds/ellie_sound10.wav.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: da729123cb5d8e24083dc3c77e88b10f
+guid: 5e2a80afaa40f4347b9256ed0ce423cc
 AudioImporter:
   externalObjects: {}
   serializedVersion: 6
