+++ Assets/Resources/Sounds/ellie_sound2.wav.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: de81f7853070da24bb706637b7242316
+guid: 3734291807cf828408cf2ba258684cb2
 AudioImporter:
   externalObjects: {}
   serializedVersion: 6
