+++ Assets/Resources/Sounds/ellie_sound3.wav.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e262a4637857ab043b9edc3d05b5f4bd
+guid: 4671e18a5aa078b438b3c81fd83afc1c
 AudioImporter:
   externalObjects: {}
   serializedVersion: 6
