+++ Assets/Resources/Sounds/ellie_sound4.wav.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4021e277e8f3b0746a2d33c4be5051d1
+guid: ae9d473a0d8036d49b2294b53b50e22d
 AudioImporter:
   externalObjects: {}
   serializedVersion: 6
