+++ Assets/Resources/Sounds/ellie_sound5.wav.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a59b069fa1eacdd4db9d4e5f7e2aff65
+guid: 635330c336291984c9e5887821f5c885
 AudioImporter:
   externalObjects: {}
   serializedVersion: 6
