+++ Assets/Resources/Sounds/ellie_sound6.wav.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d2d7f9b19b0aa5b47977a9b95af21c66
+guid: e5b1a91be260e4c4ca3d7dc69ed4525e
 AudioImporter:
   externalObjects: {}
   serializedVersion: 6
