+++ Assets/Resources/Sounds/ellie_sound7.wav.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2224df5937504b6478b42bcee7b17b2e
+guid: 70a57a24b45668643b703612cbcbd8e8
 AudioImporter:
   externalObjects: {}
   serializedVersion: 6
