+++ Assets/Resources/Sounds/ellie_sound8.wav.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8edd4a70b1c19b348bae45cb470486c9
+guid: ecc93896a4ec9a84c8bfdd645c150d61
 AudioImporter:
   externalObjects: {}
   serializedVersion: 6
