+++ Assets/Resources/Sounds/ellie_sound9.wav.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5806e22edf8ea4946934a7a0a9bad31b
+guid: eeff90e25dd6fe5409618f8e375350d9
 AudioImporter:
   externalObjects: {}
   serializedVersion: 6
