+++ Assets/Resources/Sounds/slingshot_sound1.mp3.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 44257ad476429c24a89adf1e4f7abbaf
+guid: 9956d32cd7f678449adff8bae52ab052
 AudioImporter:
   externalObjects: {}
   serializedVersion: 6
