+++ Assets/Resources/Sounds/slingshot_sound2.mp3.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4d83116111537a1498d6fae0648a2f60
+guid: cd9c19417ae5d16498a3cb2f77b21867
 AudioImporter:
   externalObjects: {}
   serializedVersion: 6
