+++ Assets/Resources/Sounds/slingshot_sound3.mp3.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 08a9dc6841882c346bb93b2aab6e4885
+guid: c07468064ec43b44f9d7aec2079e2dc3
 AudioImporter:
   externalObjects: {}
   serializedVersion: 6
