+++ Assets/Resources/Sounds/slingshot_sound4.mp3.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2500df04b6529c145b7de37329344f86
+guid: 6d134cbf0b52c7b43a345ed9b484c739
 AudioImporter:
   externalObjects: {}
   serializedVersion: 6
