+++ Assets/Scripts/UI/Loading/LoadingUI.cs
@@ -29,7 +29,7 @@ public class LoadingUI : MonoBehaviour
     private void Start()
     {
         UpdateImageTip();
-        StartCoroutine(LoadLevelAsync(SceneCenter.Instance.CurrentScene));
+        StartCoroutine(LoadLevelAsync(SceneLoadManager.Instance.CurrentScene));
     }
 
     void UpdateImageTip()
@@ -89,7 +89,7 @@ public class LoadingUI : MonoBehaviour
         yield return new WaitForSeconds(spareTimeToLoad);
 
         loadingSlider.value = 0.0f;
-        SceneCenter.Instance.FinishLoading();
+        SceneLoadManager.Instance.FinishLoading();
 
         SceneManager.UnloadSceneAsync((int)SceneName.LoadingScene);
     }
